Initial work on logging game actions
Note that the actions refer to entities as instances, rather than entities that are currently in any of the game zones; this is because the action may remove the entity from the game zone, therefore making the battlefield (e.g.) reference invalid.

This could make it difficult in the future to implement AI etc, maybe - but for now, we'll just assume all Actions are permanent and non-reversible.