Replace collection scopes with helper methods
This is because since our scopes use arrays rather than ActiveRecord,
they are not refreshed when the underlying collection is refreshed,
causing weird cache errors.

There's probably a way to resolve this with ActiveRecord trickery/callbacks,
but this seems too ugly, so we just add some simple helper methods instead.