Make tapping part of the cost and effect of activated abilities
Rather than forcing a user to tap a card, and then pay the costs -
which meant that it was extremely difficult to keep track of when a card
had been tapped and how (an effect, a user etc) - we instead assert
that the cost of playing an activated ability _includes_ that the card
can be tapped. Therefore, activating the ability taps the card.